+++ examples/surreal_ql/ql_delete_examples.py
@@ -5,11 +5,11 @@ from surrealist import Database, RecordId
 
 # Notice: all queries below not executed, just generate representation.
 # To run it against SurrealDB, you need to use run method
-with Database("http://127.0.0.1:8000", 'test', 'test', credentials=("root", "root")) as db:
+with Database("http://127.0.0.1:8000", 'test', 'test', credentials=("user_db", "user_db")) as db:
     print(db.person.delete())  # DELETE person;
-    print(db.person.delete("tobie"))  # DELETE person:tobie;
-    print(db.person.delete(RecordId("person:tobie")))  # DELETE person:tobie;
-    print(db.table("person").delete("tobie").only())  # DELETE ONLY person:tobie;
+    tobie = RecordId("tobie", table="person")  # Pay attention how you should use RecordId
+    print(db.person.delete(tobie))  # DELETE person:tobie;
+    print(db.table("person").delete(tobie).only())  # DELETE ONLY person:tobie;
 
     # DELETE person WHERE age < 18 RETURN NONE TIMEOUT 10s PARALLEL;
     print(db.table("person").delete().where("age < 18").return_none().timeout("10s").parallel())